Factor out transfer_job_s cleanup
We were cleaning up transfer_job_s in multiple places, which is the sort of
thing that leads to bugs down the road. Factored out alloc/dealloc and DMA
setupup/teardown into pairs of functions.

Also took submitting transfers to the hardware out of create_descriptor_list,
as I didn't expect a function with that name to actually submit transfers.